requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0

# API and networking
requests>=2.31.0

# Date/time handling
python-dateutil>=2.8.0

# Logging and debugging
loguru>=0.7.0
